netmiko/requirements.txt
pylint
netmiko
jinja2
pyyaml
pprint
